+++ src/test/java/nl/_42/restsecure/autoconfigure/authentication/DefaultAuthenticationResultProviderTest.java
@@ -1,6 +1,9 @@
 package nl._42.restsecure.autoconfigure.authentication;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
@@ -13,7 +16,16 @@ class DefaultAuthenticationResultProviderTest {
         User user = new User("henk", "admin");
 
         AuthenticationResult result = provider.toResult(null, null, user);
+        assertTrue(result.isAuthenticated());
         assertEquals("henk", result.getUsername());
         assertEquals(user.getAuthorities(), result.getAuthorities());
+    }
+
+    @Test
+    void to_result_whenUserIsNull_shouldReturnAnonymousResult() {
+        AuthenticationResult result = provider.toResult(null, null, null);
+        assertFalse(result.isAuthenticated());
+        assertNull(result.getUsername());
+        assertTrue(result.getAuthorities().isEmpty());
     }
 }